requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
expiringdict==1.2.1
idna==2.10
multidict==4.7.6
pydantic==1.7.2
python-qbittorrent==0.4.2
requests==2.25.0
six==1.15.0
soupsieve==2.0.1
transmission-clutch==4.2.1
urllib3==1.26.2
yarl==1.5.1
